Fixing Docker image build and publish
- added iac files to create the ECR repository
- fixed build-image.sh to correctly build a Docker image
- configured scripts to accept some metadata to be used by the created Docker image as build-args

+++ util_scripts/build-image.sh
@@ -1,4 +1,4 @@
-#! /bin/bash -e
+#! /bin/bash -ex
 
 function usage {
     cat <<EOF
@@ -9,7 +9,7 @@ DESCRIPTION:
     This script builds the hellohippo app docker image and publishes it to AWS ECR.
 
 USAGE:
-    $0 AWS_PROFILE VERSION REPOSITORY_NAME
+    $0 AWS_PROFILE VERSION REPOSITORY_NAME COMMIT
 
 PARAMETERS:
     AWS_PROFILE: Refers to an existing AWS PROFILE tied to a specific account.
@@ -18,6 +18,8 @@ PARAMETERS:
              the application version.
     
     REPOSITORY_NAME: Refers to the ECR repository NAME in the specified AWS account.
+
+    COMMIT: Refers to git commit sha used to build this version.
 =============================================================================================
 EOF
 }
@@ -27,6 +29,7 @@ EOF
 AWS_PROFILE="$1"
 VERSION="$2"
 REPOSITORY_NAME="$3"
+COMMIT="$4"
 
 if [ -z "$AWS_PROFILE" ]; then
     echo -e  "[ERROR]: No AWS_PROFILE was specified.\n"
@@ -46,8 +49,11 @@ if [ -z "$REPOSITORY_NAME" ]; then
     exit 3
 fi
 
-# Logic starts here
-ORIGINAL_DIR=`pwd`
+if [ -z "$COMMIT" ]; then
+    echo -e  "[ERROR]: No COMMIT was specified.\n"
+    usage
+    exit 4
+fi
 
 if [ -d /tmp/hellohippo/golang-app ]; then
     rm -rf /tmp/hellohippo/golang-app
@@ -59,21 +65,21 @@ git clone --depth 1 --single-branch --no-tags -b main \
 
 cd /tmp/hellohippo/golang-app
 
-REPOSITORY_URI=`aws --profile "$AWS_PROFILE" \
-    --query 'repositories[?contains(repositoryName, `my`)].repositoryUri' \
+REPOSITORY_URI=$(aws --profile "$AWS_PROFILE" \
+    --query 'repositories[?contains(repositoryName, `golang`)].repositoryUri' \
     --output text \
-    ecr describe-repositories`
+    ecr describe-repositories)
     
 IMAGE="${REPOSITORY_URI}/golang-app:$VERSION"
 
 docker buildx build \
-    --build-context src="/tmp/hellohippo/golang-app/golang-app \
+    --build-arg VERSION="$VERSION" \
+    --build-arg COMMIT="$COMMIT" \
+    --build-context src="/tmp/hellohippo/golang-app/golang-app" \
     -t  "$IMAGE" \
-    .
+    /tmp/hellohippo/golang-app/golang-app/docker
 
 aws --profile "$AWS_PROFILE" ecr get-login-password | \
     docker login -u AWS --password-stdin
 
 docker push "${IMAGE}"
-
-cd "$ORIGINAL_DIR"
